Adapt C++ readout interface.
Change to hard coded file path as in the case of the python code. This
removes the need to give the file path as argument to the constructor
class. And allows for switch between embedded and MC scale factors by
boolean flag. The specific methods for embedded scale factors are now
obsolete and are removed.