Add efficiencies for the VBF ditau trigger.
The efficiencies are added for the 2017 and 2018 data-taking periods
where the trigger has been active.
The efficiencies are added to the existing input root files.
The readout is adapted to read the efficiencies for the combined three
prong decay modes.
The test are updated to print the VBF ditau trigger efficiencies as
well.